--- Assets/Scripts/Runtime/CH1/Title/TitleController.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-namespace Runtime.CH1.Title
-{
-    public class TitleController : MonoBehaviour
-    {
-        
-    }
-}
