--- Assets/Scripts/Runtime/CH1/Title/TitleController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5db460791d3b7b34ea1bc695d36ace44
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
